pyproject.toml
[project]
name = "discordbot"
version = "0.0.1"
dependencies = [
    "discord.py >= 2.1.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"